Update sort_strings
min_i can be 0 from the start instead of None, because
after the while-loop test we already know that the list
has at least 1 element. If it has only one and no more
then min_i will be first and single element (index 0).

This way we can also start the for-loop comparisons from
index 1 instead of 0.