Update get_line_number_by_title
Not breaking or returning from the for-loop's body immediately when the
first line with the given title is found is an error and it's
unnecessary.

If we don't return immediately when the first title is found then the
iteration will move forward in the list and we'll potentially overwrite
the line_number with another value if the same title is found again
(this can happen, since the function should work any kind of input with
the same format as the game_stat.txt - e.g when there the file has
duplicate lines).

The line_number variable is unnecessary in this case and we can just use
a return statement instead. This way the if condition can be omitted
too.